LayoutTests/http/tests/cache/disk-cache/disk-cache-404-status-code.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var tests =
[
 { responseHeaders: {'Status': '404', 'Cache-control': 'max-age=0' } },
 { responseHeaders: {'Status': '404', 'Cache-control': 'max-age=100' } },
 { responseHeaders: {'Status': '404', 'Last-Modified': 'now(-1000)' } }, // Heuristic freshness.
];

description("Test that responses with HTTP status code 404 are cacheable");

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
