LayoutTests/http/tests/cache/disk-cache/disk-cache-404-status-code-expected.txt

Test that responses with HTTP status code 404 are cacheable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 3 tests

response headers: {"Status":"404","Cache-control":"max-age=0"}
response source: Network

response headers: {"Status":"404","Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Status":"404","Last-Modified":"now(-1000)"}
response source: Disk cache

PASS successfullyParsed is true

TEST COMPLETE
